chore(analysisrun): rename
